Extract matches handlers to own folder and module
