+++ src/formatting.js
@@ -15,19 +15,8 @@ function formatGameName(gameName) {
 
 const formatUser = ({ challongeUsername, challongeEmailHash }) => `*${challongeUsername}* (${challongeEmailHash ? 'verified' : 'unverified'})`;
 
-const formatMatch = (match, currentUserEmailHash) =>
-    `${formatParticipant(match.player1, currentUserEmailHash)} vs ${formatParticipant(match.player2, currentUserEmailHash)}`;
-
-const formatParticipant = (participant, currentUserEmailHash) => participant
-    ? participant.email_hash === currentUserEmailHash
-        ? `*${participant.display_name}* (you)`
-        : `*${participant.display_name}*`
-    : ':grey_question:';
-
 module.exports = {
     formatTimestamp,
     formatGameName,
     formatUser,
-    formatMatch,
-    formatParticipant,
 };
